+++ src/test/testUtils/index.ts
@@ -5,4 +5,6 @@ export * from "./abridgedTestUtils";
 export * from "./analysisTestUtils";
 export * from "./assertEqualTestUtils";
 export * from "./inspectionTestUtils";
+export * from "./parseTestUtils";
 export * from "./testUtils";
+export * from "./validationTestUtils";
